add jpg quali check SHOP-4523

+++ admin/includes/admin_tools.php
@@ -1,5 +1,6 @@
 <?php
 
+use JTL\Alert\Alert;
 use JTL\Backend\AdminFavorite;
 use JTL\Backend\Notification;
 use JTL\Campaign;
@@ -221,6 +222,7 @@ function saveAdminSectionSettings(int $configSectionID, array &$post, $tags = [C
         return __('errorCSRF');
     }
     $db       = Shop::Container()->getDB();
+    $invalid  = 0;
     $confData = $db->selectAll(
         'teinstellungenconf',
         ['kEinstellungenSektion', 'cConf'],
@@ -228,14 +230,12 @@ function saveAdminSectionSettings(int $configSectionID, array &$post, $tags = [C
         '*',
         'nSort'
     );
-    if (count($confData) === 0) {
-        return __('errorConfigSave');
-    }
     foreach ($confData as $config) {
         $val                        = new stdClass();
         $val->cWert                 = $post[$config->cWertName] ?? null;
         $val->cName                 = $config->cWertName;
         $val->kEinstellungenSektion = $configSectionID;
+        $valid                      = true;
         switch ($config->cInputTyp) {
             case 'kommazahl':
                 $val->cWert = (float)str_replace(',', '.', $val->cWert);
@@ -243,6 +243,7 @@ function saveAdminSectionSettings(int $configSectionID, array &$post, $tags = [C
             case 'zahl':
             case 'number':
                 $val->cWert = (int)$val->cWert;
+                $valid      = validateSetting($val);
                 break;
             case 'text':
                 $val->cWert = mb_substr($val->cWert, 0, 255);
@@ -253,7 +254,7 @@ function saveAdminSectionSettings(int $configSectionID, array &$post, $tags = [C
                 break;
         }
 
-        if ($config->cInputTyp !== 'listbox' && $config->cInputTyp !== 'selectkdngrp') {
+        if ($valid && $config->cInputTyp !== 'listbox' && $config->cInputTyp !== 'selectkdngrp') {
             $db->delete(
                 'teinstellungen',
                 ['kEinstellungenSektion', 'cName'],
@@ -261,10 +262,56 @@ function saveAdminSectionSettings(int $configSectionID, array &$post, $tags = [C
             );
             $db->insert('teinstellungen', $val);
         }
+        if (!$valid) {
+            $invalid++;
+        }
     }
     Shop::Container()->getCache()->flushTags($tags);
 
+    if ($invalid > 0 || \count($confData) === 0) {
+        return __('errorConfigSave');
+    }
+
     return __('successConfigSave');
+}
+
+/**
+ * @param $setting
+ * @return bool
+ */
+function validateSetting($setting)
+{
+    $valid = true;
+    switch ($setting->cName) {
+        case 'bilder_jpg_quali':
+            $valid = validateNumberRange(0, 100, $setting);
+            break;
+        default:
+            break;
+    }
+
+    return $valid;
+}
+
+/**
+ * @param int $min
+ * @param int $max
+ * @param $setting
+ * @return bool
+ */
+function validateNumberRange(int $min, int $max, $setting)
+{
+    $valid = $min <= $setting->cWert && $setting->cWert <= $max;
+
+    if (!$valid) {
+        Shop::Container()->getAlertService()->addAlert(
+            Alert::TYPE_DANGER,
+            'number falsch',
+            'errrorNumberRange'
+        );
+    }
+
+    return $valid;
 }
 
 /**
